requirements.txt
# --- Web & API Core ---
Flask==3.1.2
Werkzeug==3.1.5
Jinja2==3.1.6
MarkupSafe==3.0.3
itsdangerous==2.2.0
click==8.3.1
blinker==1.9.0

# --- HTTP & Networking ---
requests==2.32.5
httpx==0.28.1
httpcore==1.0.9
httplib2==0.31.2
urllib3==2.6.3
h11==0.16.0
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11

# --- AI, Speech & Machine Learning ---
torch==2.9.1
triton==3.5.1
numpy==2.4.2
onnxruntime==1.24.1
faster-whisper==1.2.1
whisper-s2t-reborn==1.4.4
ctranslate2==4.6.2
tokenizers==0.22.2
huggingface_hub==1.4.1
hf-xet==1.2.0
av==16.1.0

# --- NVIDIA CUDA Runtime (CU12) ---
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90

# --- Google Cloud & APIs ---
google-api-python-client==2.189.0
google-api-core==2.29.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
googleapis-common-protos==1.72.0
proto-plus==1.27.1
protobuf==6.33.5

# --- Scraping & Media ---
yt-dlp==2026.2.4
beautifulsoup4==4.14.3
soupsieve==2.8.3
lxml==6.0.2

# --- Utilities & Math ---
cryptography==46.0.4
pyasn1==0.6.2
pyasn1_modules==0.4.2
rsa==4.9.1
psycopg2-binary==2.9.11
PyYAML==6.0.3
tqdm==4.67.3
filelock==3.20.3
fsspec==2026.2.0
packaging==26.0
platformdirs==4.5.1
shellingham==1.5.4
typer-slim==0.21.1
anyio==4.12.1
typing_extensions==4.15.0
sympy==1.14.0
mpmath==1.3.0
networkx==3.6.1
flatbuffers==25.12.19
uritemplate==4.2.0
pycparser==3.0
pyparsing==3.3.2
python-dotenv==1.2.1